ingress-api/requirements-db.txt
sqlalchemy
psycopg2-binary
alembic
python-dotenv
